Add files via upload
temporary build of a web app that intakes a list of ingredients and returns a list of potential recipes that can be completed with the following ingredient. Works by scraping food.com searching for the ingredients needed. 

The only issue that I need to figure out as of now is the concurrent scraping, as Im implementing this to make the scraping much faster, but I keep getting a type error "most recent call last". Will fix tomo.