Add files via upload
This is a test space for a simple blockchain that I am developing through python. It is not intended for commercial use, simply for a better understanding of WEB3 and the technology associated with it. This is an attempt to create a 'private blockchain' where the blocks that are located in the ledger are non-fungible tokens, saved as images. The idea is that you can add a timestamped image to this immutable chain and that this image is then associated with a specific machine that entered it. This then can confirm and verify who created an image, and when that image was created, indefinitely. You might see why this would have good real-world use.

Current Version: Blockchain itself works and uses an SHA to generate the hash codes for each block. The CLI is still a little 'wonky' as I have to fix some typo errors, I want to commit this as of now, but I will fix the typos shortly. I wrote the CLI program on a separate machine and forgot how I formatted the names of the methods and such in the origional blockchain program, thus I will go back and correct this soon.